FastAPI/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.3
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
distro==1.9.0
fastapi==0.115.12
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.9.0
openai==1.70.0
outcome==1.3.0.post0
packaging==24.2
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.11.1
pydantic_core==2.33.0
PySocks==1.7.1
python-dotenv==1.1.0
python-multipart==0.0.20
requests==2.32.3
selenium==4.30.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
starlette==0.46.1
tqdm==4.67.1
trio==0.29.0
trio-websocket==0.12.2
typing-inspection==0.4.0
typing_extensions==4.13.0
urllib3==2.3.0
uvicorn==0.34.0
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
